#326: move CLI11 inside a subfolder to avoid clashes with VT that is system installed
